rename the files to remove "other-modes" from the filenames
