+++ source/views/transportation/other-modes/index.js
@@ -1,4 +1,4 @@
 // @flow
 
-export {OtherModesView} from './other-modes-list'
-export {OtherModesDetailView} from './other-modes-detail'
+export {OtherModesView} from './list'
+export {OtherModesDetailView} from './detail'
